Added SVG Export to pySlice.py
The main python script now outputs all of the pairs of points as lines
in an SVG file. This is still incomplete, as:
We don't have Python 3 support
It doesn't take into account triangles that lie on a plane #3 (all the
same Z coordinates), so that is why we still end up with missing gaps.
This is getting towards fixing #2